+++ Migrations/20200116125925_InitialCreate.cs
@@ -72,7 +72,7 @@ namespace heal_fit.Migrations
                 {
                     ID = table.Column<int>(nullable: false)
                         .Annotation("Sqlite:Autoincrement", true),
-                    Value = table.Column<float>(nullable: false),
+                    Value = table.Column<double>(nullable: false),
                     Type = table.Column<int>(nullable: false),
                     Date = table.Column<DateTime>(nullable: false),
                     ProfileID = table.Column<int>(nullable: false)
